refactor: separate factory pattern into handler.py
