+++ src/main.py
@@ -1,48 +1,14 @@
 import sys
-from abc import ABC, abstractmethod
 
 from config.settings import PROJECT_ROOT, settings
 from modules.intent import IntentModule
-from modules.compare import CompareModule
-from modules.policy import PolicyModule
 from db.sql_utils import TemplateManager
-from options.enums import IntentType
+from modules.handler import HandlerFactory
 
 from openai import OpenAI
 
 
 # TODO: isort setting precommit setting
-class Handler(ABC):
-    @abstractmethod
-    def handle(self, question: str) -> str:
-        pass
-
-
-class CompareHandler(Handler):
-    def __init__(self, llm_client: OpenAI, template_manager: TemplateManager):
-        self.compare_module = CompareModule(llm_client, template_manager)
-
-    def handle(self, question: str) -> str:
-        return self.compare_module.get_search_result(question)
-
-
-class PolicyHandler(Handler):
-    def __init__(self, llm_client: OpenAI, template_manager: TemplateManager):
-        self.policy_module = PolicyModule(llm_client, template_manager)
-
-    def handle(self, question: str) -> str:
-        return self.policy_module.respond_policy_question(question)
-
-
-class HandlerFactory:
-    @staticmethod
-    def get_handler(
-        intent: str, llm_client: OpenAI, template_manager: TemplateManager
-    ) -> Handler:
-        if intent == IntentType.COMPARE_QUESTION:
-            return CompareHandler(llm_client, template_manager)
-        return PolicyHandler(llm_client, template_manager)
-
 
 if __name__ == "__main__":
     print("\n=== 보험 상담 챗봇 (Factory Pattern) ===")
